requirements.txt
pip==21.1.1
numpy==1.19.2
matplotlib==3.4.1
pandas==1.1.3
streamlit==0.81.1
plotly==4.14.3 
scikit-learn==0.24.2
tensorflow-cpu==2.4.1
statsmodels==0.12.2